Replaced paths file with a file that bootstraps all the others
Issues with symlinks causing the same file to init as different things(see desc)

Had issues with loading packages from init, as when it is first called as the system init
via symlink it installs all packages setting the path to the init symlink name(.emacs.d)
and the user set home dir (C:/repo/.../), which doesn't exist.  Then on next boot the
user-init-file is used(~/dot-emacs/)  so when use-package looks for packages it finds
none if them as the path it believes they are on doesn't exist, so it creates this path
and reinstalls.  Fixed this by making .emacs the system init, which then calls the
user-init-file, from the set directories which then installs with use-package.

TDLR; Really comfusing issue